requirements.txt
Django==3.2.5
django-bootstrap5==2.1.1
django-filter==2.4.0
django-tables2==2.4.0
